Fix the clear bit method and move missile collision to player

+++ test/MissileSpec.hs
@@ -1,10 +1,8 @@
 module MissileSpec where
 
-import Missile
-
 import Test.Hspec
 
 spec :: Spec
 spec = do
   describe "moveMissile" $ do
-    it "should be implemented" $ 1 `shouldBe` 1
+    it "should be implemented" $ (1::Int) `shouldBe` 1
