+++ test/ObjectiveSpec.hs
@@ -1,10 +1,8 @@
 module ObjectiveSpec where
 
-import Objective
-
 import Test.Hspec
 
 spec :: Spec
 spec = do
   describe "myBoardScore" $ do
-    it "should be implemented" $ 1 `shouldBe` 1
+    it "should be implemented" $ (1::Int) `shouldBe` 1
